change shadow memory type from uint8* to int*

+++ include/analysis/memory_heatmap.h
@@ -87,7 +87,7 @@ Map<u64, std::shared_ptr<Memory>> _sub_memories;
  */
 struct HeatMapMemory {
   size_t size;
-  uint8_t *array;
+  int *array;
 
   HeatMapMemory() = default;
